changed file extensions to cpp
