+++ build.sh
@@ -1,6 +1,8 @@
 # Create an executable
 cd build
 
+rm -rf ./*
+
 cmake ..
 
 make
