--- src/structs/student.h
@@ -1,9 +1,0 @@
-
-typedef struct student {
-    char *first_name;
-    char *last_name;
-    char *student_id;
-    int col;
-    int row;
-    struct student *next;
-} student;
